fix #55: 조회 수 캐시 오류 수정

+++ src/main/java/com/codesquad/secondhand/item/infrastructure/ItemCustomRepository.java
@@ -22,7 +22,7 @@ public interface ItemCustomRepository {
 
 	Optional<ItemDetailDto> findDetailById(Long id);
 
-	Long incrementViewCount(Long id);
+	int incrementViewCount(Long id);
 
 	Map<Long, Integer> findAllRedisItemViewCount();
 }
